// src/enums/enums.ts
export enum FilterState {
  ALL = "All",
  COMPLETED = "Completed",
  PENDING = "Pending",
}
export enum Tab {
  DASHBOARD = "Dashboard",
  MY_TASKS = "MyTasks",
  TASKS = "Tasks",
  HISTORY = "History",
}

export enum Repeat {
  MONTHLY = "Monthly",
  WEEKLY = "Weekly",
  YEARLY = "Yearly",
}

export enum Priority {
  URGENT = "Urgent",
  LOW = "Low",
  MEDIUM = "Medium",
}